+++ jobsummaries/jobsummary_test.go
@@ -2,13 +2,13 @@ package jobsummaries
 
 import (
 	buildinfo "github.com/jfrog/build-info-go/entities"
-	"github.com/jfrog/jfrog-cli-core/v2/artifactory/commands/githubsummariesimpl"
+	"github.com/jfrog/jfrog-cli-core/v2/artifactory/commands/jobsummariesimpl"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
 func TestBuildUiUrl(t *testing.T) {
-	gh := &githubsummariesimpl.GithubSummaryRtUploadImpl{
+	gh := &jobsummariesimpl.GithubSummaryRtUploadImpl{
 		PlatformUrl:     "https://myplatform.com/",
 		JfrogProjectKey: "myProject",
 	}
@@ -16,7 +16,7 @@ func TestBuildUiUrl(t *testing.T) {
 	actual := gh.BuildUiUrl("myPath")
 	assert.Equal(t, expected, actual)
 
-	gh = &githubsummariesimpl.GithubSummaryRtUploadImpl{
+	gh = &jobsummariesimpl.GithubSummaryRtUploadImpl{
 		PlatformUrl:     "https://myplatform.com/",
 		JfrogProjectKey: "",
 	}
@@ -26,7 +26,7 @@ func TestBuildUiUrl(t *testing.T) {
 }
 
 func TestBuildInfoTable(t *testing.T) {
-	gh := &githubsummariesimpl.GithubSummaryBpImpl{}
+	gh := &jobsummariesimpl.GithubSummaryBpImpl{}
 	gh.Builds = []*buildinfo.BuildInfo{
 		{
 			Name:     "buildName",
@@ -41,10 +41,10 @@ func TestBuildInfoTable(t *testing.T) {
 
 func TestParseBuildTime(t *testing.T) {
 	expected := "Jan 2, 2006 15:04:05"
-	actual := githubsummariesimpl.ParseBuildTime("2006-01-02T15:04:05.000-0700")
+	actual := jobsummariesimpl.ParseBuildTime("2006-01-02T15:04:05.000-0700")
 	assert.Equal(t, expected, actual)
 
 	expected = "N/A"
-	actual = githubsummariesimpl.ParseBuildTime("")
+	actual = jobsummariesimpl.ParseBuildTime("")
 	assert.Equal(t, expected, actual)
 }
